requirements.txt
# Core - versions fixes pour éviter les conflits
python-dotenv==1.1.0
tqdm==4.67.1
pillow==11.2.1
plotly==6.0.1

# PDF & multimodal
PyMuPDF==1.26.0
unstructured[pdf]==0.17.2
pytesseract==0.3.13

# Streamlit interface
streamlit==1.45.1

# LLM & vecteurs - versions compatibles
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.23
langchain-google-genai==2.0.10
google-generativeai==0.8.5
chromadb==1.0.12

# GPT4All - optionnel, peut être retiré si non utilisé
gpt4all==2.0.2

# Test
pytest==8.3.5